metro/requirements.txt
pyserial
bitstring
redis
rq
